Send Project to AudioEngine instead of SystemState
* Move mixer and sampler to `Project`, so the audio engine can safely
  copy `Project` between threads.
* Change audio devices using `System` rather than the UI Event Queue
  (also removing Settings events from UI Event enum)
* Minimize duplicated code, clean up variable names and structs

+++ src/data/mixer/channel.rs
@@ -1,5 +1,5 @@
 pub struct Channel {
-	pub volume: f32,
+	pub volume: f64,
 	pub panning: f32,
 	pub muted: bool,
 	pub name: String,
